adding player module based on a service
