+++ app/build.gradle
@@ -30,7 +30,7 @@ android {
 }
 
 dependencies {
-    implementation project(":no-player")
+    implementation project(":player")
 
     implementation libraries.kotlinStdLib
     implementation libraries.androidxAppCompat
