+++ player/build.gradle
@@ -1,0 +1,17 @@
+apply plugin: 'com.android.library'
+apply plugin: 'kotlin-android'
+
+android {
+  compileSdkVersion versions.androidSdk.compile
+
+  defaultConfig {
+    minSdkVersion versions.androidSdk.min
+    targetSdkVersion versions.androidSdk.target
+  }
+}
+
+dependencies {
+  implementation project(":no-player")
+  implementation libraries.androidxAppCompat
+  implementation libraries.kotlinStdLib
+}
