+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', 'no-player'
+include ':app', ':player', 'no-player'
